+++ tencentcloud/cme/v20191029/client.go
@@ -618,7 +618,7 @@ func NewDescribeTeamsResponse() (response *DescribeTeamsResponse) {
     return
 }
 
-// 获取指定团队的信息。
+// 获取指定团队的信息，拉取团队信息列表。
 func (c *Client) DescribeTeams(request *DescribeTeamsRequest) (response *DescribeTeamsResponse, err error) {
     if request == nil {
         request = NewDescribeTeamsRequest()
@@ -793,7 +793,7 @@ func NewImportMaterialResponse() (response *ImportMaterialResponse) {
     return
 }
 
-// 将云点播媒资文件导入到云剪素材库。
+// 将云点播媒资文件导入到云剪媒体资源库。
 func (c *Client) ImportMaterial(request *ImportMaterialRequest) (response *ImportMaterialResponse, err error) {
     if request == nil {
         request = NewImportMaterialRequest()
@@ -818,7 +818,7 @@ func NewImportMediaToProjectResponse() (response *ImportMediaToProjectResponse) 
     return
 }
 
-// 将云点播中的媒资添加到素材库中，供后续视频编辑使用。
+// 将云点播中的媒资或者用户自有媒资文件添加到媒体库中，供后续视频编辑使用。
 func (c *Client) ImportMediaToProject(request *ImportMediaToProjectRequest) (response *ImportMediaToProjectResponse, err error) {
     if request == nil {
         request = NewImportMediaToProjectRequest()