+++ tencentcloud/faceid/v20180301/client.go
@@ -408,7 +408,7 @@ func NewGetRealNameAuthTokenResponse() (response *GetRealNameAuthTokenResponse) 
     return
 }
 
-// 输入姓名和身份证号验证与微信实名的信息是否一致
+// 该接口仅限微信公众号中使用，传入姓名和身份证号获取回调URL，在微信公众号中打开验证姓名和身份证号与微信实名的信息是否一致。
 func (c *Client) GetRealNameAuthToken(request *GetRealNameAuthTokenRequest) (response *GetRealNameAuthTokenResponse, err error) {
     if request == nil {
         request = NewGetRealNameAuthTokenRequest()